requirements.txt
streamlit==0.86.0
numpy==1.19.2
plotly==5.1.0
joblib==0.17.0
scikit-learn==0.23.2
pillow==8.0.1